server/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.1.13
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
idna==3.4
pymongo==3.12.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.2.4
typing_extensions==4.8.0
urllib3==2.0.7
